requirements.txt
flask==3.0.3
python-dotenv==1.0.1
gunicorn==20.0.4
Flask-Cors==4.0.1
requests==2.31.0
PyJWT==2.8.0